improve ensure nodeclaims according to review comments
